requirements.txt
langchain
langchain-astradb
langchain-community
langchain-huggingface
langchain-groq
langchain-community
datasets
pypdf
python-dotenv
pandas
flask
flask-cors
prometheus-client

langchain-classic
huggingface-hub==0.36.2
huggingface-hub==0.36.2
sentence-transformers
transformers<5